Speech to txt complete

+++ video_processing/process_videos.py
@@ -8,9 +8,69 @@ import csv
 import cv2
 import shutil
 from multiprocessing import Pool, Manager, cpu_count, Process, Queue
+#import speech_to_text as stt
 
 REQUIRED_FPS = 3
 SKIP_FRAMES = 60//REQUIRED_FPS+1
+
+import time
+import azure.cognitiveservices.speech as speechsdk
+
+# Creates an instance of a speech config with specified subscription key and service region.
+# Replace with your own subscription key and region identifier from here: https://aka.ms/speech/sdkregion
+speech_key, service_region = "d215959e99c24caeb2ee4e620016782f", "westus"
+speech_config = speechsdk.SpeechConfig(subscription=speech_key, region=service_region)
+speech_config.request_word_level_timestamps()
+
+# Creates an audio configuration that points to an audio file.
+# Replace with your own audio filename.
+def speech_to_text(audio_filename, output_filename=None):
+    print("Speech To Text conversion")
+    if not output_filename:
+        output_filename = audio_filename.split('.')[0]+f'_text.json'
+
+    data = list()
+    done = False
+
+    audio_input = speechsdk.audio.AudioConfig(filename=audio_filename)
+    # Creates a recognizer with the given settings
+    speech_recognizer = speechsdk.SpeechRecognizer(speech_config=speech_config, audio_config=audio_input)
+
+    print("Recognizing first result...")
+
+    def stop_cb(evt):
+        """callback that signals to stop continuous recognition upon receiving an event `evt`"""
+        print('CLOSING on {}'.format(evt))
+        nonlocal done, output_filename
+        f = open(output_filename, 'w')
+        f.write(json.dumps(data))
+        f.close()
+        done = True
+
+    # Connect callbacks to the events fired by the speech recognizer
+    def store_text(evt):
+        nonlocal data, output_filename
+        print(f"writing to {output_filename}")
+        #f = open(output_filename, 'w')
+        #f.write(evt.result.json)
+        #f.close()
+        data.append(evt.result.json)
+
+    speech_recognizer.recognizing.connect(lambda evt: print('RECOGNIZING: {}'.format(evt.result.json)))
+    speech_recognizer.recognized.connect(store_text)#lambda evt: print('RECOGNIZED: {}'.format(evt.result.json)))
+    speech_recognizer.session_started.connect(lambda evt: print('SESSION STARTED: {}'.format(evt)))
+    speech_recognizer.session_stopped.connect(lambda evt: print('SESSION STOPPED {}'.format(evt)))
+    speech_recognizer.canceled.connect(lambda evt: print('CANCELED {}'.format(evt)))
+    # stop continuous recognition on either session stopped or canceled events
+    speech_recognizer.session_stopped.connect(stop_cb)
+    speech_recognizer.canceled.connect(stop_cb)
+
+    # Start continuous speech recognition
+    speech_recognizer.start_continuous_recognition()
+    while not done:
+        time.sleep(.5)
+
+    return 0
 
 def get_file_name(path):
     #Get file name and remove extension
@@ -49,7 +109,7 @@ def compress_video(video_location):
         frame+=1
     return output_dir 
 
-def extract_features(video_location):
+def extract_video_features(video_location):
     '''
     Run openface on given video, read the extracted features, return the the extracted features
     '''
@@ -60,6 +120,10 @@ def extract_features(video_location):
     #features = read_features(video_location)
     return proc.returncode
 
+def extract_audio_features(audio_location):
+    output_path = 'processed_video/' + get_file_name(audio_location) + '_txt'
+    return speech_to_text(audio_location, output_path)
+    
 class ProcessVideos():
     '''
     Combined set of all videos that have to be processed
@@ -92,25 +156,41 @@ def bitrate_helper(abr):
 
     return num
 
+def webm_to_wav(inp_filepath, out_filepath):
+    args = ['ffmpeg', '-i', inp_filepath, '-ac', '1', '-f', 'wav', out_filepath]
+    proc = subprocess.run(args)
+
+
 '''
--10 video/audio failed download
--11 Compression failed
+-10video/audio failed download
+-11 Video Compression failed
+-21 Audio transformation failed
 '''
 class Video():
     '''
-    Represents a single video and features associated with it
+    Represents a single video and formats them for feature extraction
     '''
-    def __init__(self, video_id):
+    def __init__(self, video_id, process_video=False, process_audio=False):
         self.id = video_id
         self.video_location, self.audio_location = self.get_yt_video_data(video_id)
+        self.features_extracted_status = 0
+        
         if not self.video_location and not self.audio_location:
             self.features_extracted_status = -10
         else:
-            try:
-                self.frames_location = compress_video(self.video_location)
-                self.features_extracted_status = 0
-            except:
-                self.features_extracted_status = -11#Openface().extract_features(self.frames_location)
+            if process_video:
+                try:
+                    self.frames_location = compress_video(self.video_location)
+                except:
+                    self.features_extracted_status = -11#Openface().extract_features(self.frames_location)
+            if process_audio:
+                #Convert audio
+                try:
+                    self.wav_audio_location = self.audio_location.split('.')[0]+'.wav'
+                    webm_to_wav(self.audio_location, self.wav_audio_location)
+                except:
+                    self.features_extracted_status += -12
+
             #print(f"Video feature extraction for {video_id} returned with status code {self.features_extracted_status}")
     
 
@@ -120,18 +200,19 @@ class Video():
         audio_codec="opus"
         video_stream = audio_stream = None
         
+        video_location = audio_location = None
         while 1:
             try:
                 yt = YouTube(url.format(video_id))
                 print(url.format(video_id))
                 
                 #Filter available streams to get the compatible ones. Adaptive : Audio, Video downloaded separately and merged after download; progressive : Audio, Video downloaded in same file
-                adaptive_video_streams = yt.streams.filter(adaptive=True, type="video")
+                #adaptive_video_streams = yt.streams.filter(adaptive=True, type="video")
                 adaptive_audio_streams = yt.streams.filter(type="audio", audio_codec=audio_codec)
                 
                 #Download the Audio and Video into a temporary file. Assuming there is only a single file at a given time for OpenFace processing
                 #TODO Ensure the asc function works as expected for all videos. The sorting assumption is made based on the output for 4-5 test videos.
-                video_location = adaptive_video_streams.asc()[0].download(output_path="temporary_downloads/raw_video", filename=f"video_{video_id}") #asc puts the highest quality video at the top 
+                #video_location = adaptive_video_streams.asc()[0].download(output_path="temporary_downloads/raw_video", filename=f"video_{video_id}") #asc puts the highest quality video at the top 
                 audio_location = adaptive_audio_streams.asc()[-1].download(output_path="temporary_downloads/raw_audio", filename=f"audio_{video_id}")#asc puts the highest bitrate at the bottom
             except Exception as e:
                 if '429' in str(e):
@@ -151,19 +232,33 @@ class Video():
 def print_error(e):
     print(e)
 
-def extract_features_helper(video):
-    #Store id for failed videos 
-    extracted_status = extract_features(video.frames_location)
-    #for status, video_id in zip(video_proc_obj.video_status, video_ids):
-    if extracted_status!=0:
-        errors_list.append([video.id, extracted_status, 'video'])
-    else:
-        #Delete the video files
-        os.remove(video.video_location)
-        os.remove(video.audio_location)
-        shutil.rmtree(video.frames_location)
+def extract_features_helper(video, process_video=False, process_audio=False):
 
-        processed_list.append(video.id)
+    if process_video:
+        #Store id for failed videos 
+        extracted_status = extract_video_features(video.frames_location)
+        #for status, video_id in zip(video_proc_obj.video_status, video_ids):
+        if extracted_status!=0:
+            errors_list.append([video.id, extracted_status, 'video'])
+        else:
+            #Delete the video files
+            os.remove(video.video_location)
+            shutil.rmtree(video.frames_location)
+
+            processed_list.append(video.id)
+
+    if process_audio:
+        extracted_status = extract_audio_features(video.wav_audio_location)
+        #for status, video_id in zip(video_proc_obj.video_status, video_ids):
+        if extracted_status!=0:
+            errors_list.append([video.id, extracted_status, 'audio'])
+        else:
+            #Delete the video files
+            os.remove(video.wav_audio_location)
+            os.remove(video.audio_location)
+
+            processed_list.append(video.id)
+
     
     ###############################
     #Audio processing
@@ -180,7 +275,7 @@ def process_videos_helper():
     while not video_ids.empty():
         video_id = video_ids.get()
         print("Helper", video_ids)
-        video = Video(video_id)
+        video = Video(video_id, process_audio=True)
         
         #Store id for failed videos
         status = video.features_extracted_status
@@ -189,6 +284,7 @@ def process_videos_helper():
             errors_list.append([video_id, status, 'video'])
         else:
             downloaded_que.put(video)
+        #break
 
             #videos_failed_csv.writerow([video_id, status])
     
@@ -206,14 +302,30 @@ def process_videos_helper():
         '''
 
 def main():
+    #return
     #global video_ids
     #Read raw data
-    f = open('processed_video_ids.json', 'r')
+    f = open('final_ids.json', 'r')
     processed_list.extend(json.load(f))
     f.close()
 
+    f = open('processed_txt_ids.json', 'r')
+    ids_to_ignore = json.load(f)
+    f.close()
+
+    i = 0
+    for _, _id in enumerate(processed_list):
+        if _id in ids_to_ignore:
+            continue
+        if i==2000:
+            break
+        i+=1
+        video_ids.put(_id)
+        #print("adding id ", _id)
+        #break
     #video_ids = list()
-    
+   
+    '''
     f = open('Data/raw_data.csv')
     c = csv.reader(f)
     #Get rid of the header row
@@ -227,7 +339,7 @@ def main():
         if x>=7:
             break
     f.close()
-    
+    '''
     #Fetch details of all videos
     #video_ids = video_ids[:1]+['ewfherher']#, 'wO0a2C6PqtY']
     start = time.time()
@@ -240,9 +352,11 @@ def main():
 
     download_video_proc = Process(target=process_videos_helper)
     download_video_proc.start()
+    time.sleep(10)
     
     openface_processes = list()
-    MAX_PROCESSES = cpu_count()-1
+    MAX_PROCESSES = 20#cpu_count()-1
+
     while not video_ids.empty() or not downloaded_que.empty() or openface_processes:
         print("downloading ", video_ids, video_ids.empty(), downloaded_que.empty())
         time.sleep(2)
@@ -255,7 +369,7 @@ def main():
             #Get running procs
             #If running procs<max, spawn, else continue
             video = downloaded_que.get()
-            p = Process(target=extract_features_helper, args=(video,))
+            p = Process(target=extract_features_helper, args=(video, False, True,))
             p.start()
             openface_processes.append(p)
             print("Vid ID", video.id)
@@ -263,7 +377,8 @@ def main():
 
     ###################################
 
-    print("downloading ", video_ids, video_ids.empty(), downloaded_que.empty())
+    openface_processes = [p for p in openface_processes if p.is_alive()]
+    print("downloading ", video_ids, video_ids.empty(), downloaded_que.empty(), openface_processes)
     print("Processing time = ", time.time()-start)
     print("Saving processed video ids and Terminating program !!!")
     f = open('processed_video_ids.json', 'w')
